A Package for Fluvial Geomorphic Analysis

This package contains a wide range of functions for performing fluvial geomorphic analysis. This package is designed for use with the FluvialGeomorph Python package. The FluvialGeomorph package contains an ArcGIS toolbox that can be used to extract spatial data from high- resolution terrain data. FluvialGeomorph uses functions in this package to perform a wide range of fluvial geomorphic analysis (e.g., extract stream channel dimensions, help choose bankfull elevation for ungaged streams).

Details

Place some other package details here. And then some more.

Installation

The datasets used as input into this package can be created using the FluvialGeomorph-toolbox https://github.com/FluvialGeomorph/FluvialGeomorph-toolbox Python package. From within the GIS Windows application ArcGIS Pro, the arcgisbinding R package can be installed by following the instructions at https://r-arcgis.github.io/assets/arcgisbinding-vignette.html. No other functions within this package require ESRI ArcGIS Pro.

See also

Author

Maintainer: Barrie Chileen Martinez Barrie.V.Chileen@usace.army.mil (ORCID)

Authors:

Michael Dougherty Michael.P.Dougherty@usace.army.mil (ORCID)

Christopher Haring Christopher.P.Haring@usace.army.mil

Other contributors:

Thomas Darby [reviewer]

Davi Michl [reviewer]