Level 2 Report

Creates the FluvialGeomorph Level 2 report.

level_2_report(
  stream,
  flowline_fc,
  xs_fc,
  xs_dims_fc,
  xs_points_1,
  xs_points_2,
  xs_points_3,
  xs_points_4,
  survey_name_1,
  survey_name_2,
  survey_name_3,
  survey_name_4,
  dem,
  banklines_fc,
  features_fc,
  bf_estimate,
  regions,
  label_xs,
  show_xs_map = FALSE,
  profile_units,
  aerial = TRUE,
  elevation = FALSE,
  xs_label_freq = 5,
  exaggeration = 10,
  extent_factor = 1.2,
  output_dir,
  output_format
)

Arguments

stream

character; The stream name. The stream name must match a stream name in `ReachName` field in the other parameters.

flowline_fc

character; The path to a `flowline` feature class.

xs_fc

character; The path to the cross section feature class.

xs_dims_fc

character; The path to the "xs_dims" feature class.

xs_points_1

character; The path to a `xs_points` feature class for the "base year".

xs_points_2

character; The path to a `xs_points` feature class for the second time period.

xs_points_3

character; The path to a `xs_points` feature class for the third time period.

xs_points_4

character; The path to a `xs_points` feature class for the fourth time period.

survey_name_1

character: The name or date of the "base year" survey.

survey_name_2

character: The name or date of the second survey.

survey_name_3

character: The name or date of the third survey.

survey_name_4

character: The name or date of the fourth survey.

dem

character; The path to the DEM raster.

banklines_fc

character: the path to the banklines feature class.

features_fc

character; The path to a `features` feature class.

bf_estimate

numeric; Detrended bankfull estimate (units: detrended feet).

regions

character vector; Regions to calculate hydraulic dimensions for. See the `RegionalCurve` package for a list of regions.

label_xs

logical; Label cross sections?

show_xs_map

logical; Add the cross section maps to the report?

profile_units

character; the units of the longitudinal profile. One of "kilometers", "meters", "miles", or "feet".

aerial

logical; Display an overview map with an aerial photo background?

elevation

logical; Display an overview map with an elevation background?

xs_label_freq

numeric; An integer indicating the frequency of cross section labels.

exaggeration

numeric; The degree of terrain exaggeration.

extent_factor

numeric; The amount the extent is expanded around the cross section feature class. Values greater than one zoom out, values less than one zoom in.

output_dir

character; The path to the folder in which to write the report.

output_format

character; The file format of the report. One of "html_document", "word_document", "pdf_document".

Value

Produces a FluvialGeomorph Level 2 Report in the `output_dir` in the requested file format.