Reach Overview Map

Produces a reach overview map displaying cross section locations over an aerial image or elevation multi-direction shaded relief.

map_reach_overview(
  flowline_sf,
  cross_section_sf,
  background = "none",
  xs_label_freq = 1,
  exaggeration = 20,
  extent_factor = 1.1
)

Arguments

flowline_sf

sf object; A flowline feature class

cross_section_sf

sf object; A cross section feature class

background

character; The type of map background. One of "aerial", "elevation", or "none".

xs_label_freq

numeric; An integer indicating the frequency of cross section labels.

exaggeration

numeric; The degree of terrain exaggeration.

extent_factor

numeric; The amount the extent is expanded around the cross section feature class. Values greater than one zoom out, values less than one zoom in.

Value

a tmap object