Calculate the meander width for each loop

Calculates the meander width (channel belt width) for each loop in the input `bankline_points` data frame.

meander_width(bankline_points)

Arguments

bankline_points

sf data.frame; a fluvgeo bankline_points data structure

Value

Returns a data frame of loops with the calculated meander width in units feet.

Examples

bankline_points <- fluvgeo::sin_bankline_points_sf
mw <- meander_width(bankline_points = bankline_points)
#> [1] "Diagnostic report of bankline points"
#> [1] "Loop 1"
#> [1] "    Apex: TRUE   Route-M: 237.63"
#> [1] "    last loop bank: "
#> [1] "    current loop bank: right descending"
#> [1] "        Bend 1"
#> [1] "            Start: TRUE Route-M: 33.25"
#> [1] "            End: TRUE   Route-M: 296.39"
#> [1] "            bank: right descending"
#> [1] "            min_loop_m: 33.25"
#> [1] "            max_loop_m: 296.39"
#> [1] "            min_last_bend_m: 33.25"
#> [1] "            max_last_bend_m: 296.39"
#> [1] "    ** The Loop 1 apex point is located within the loop."
#> [1] "Loop 2"
#> [1] "    Apex: TRUE   Route-M: 572.41"
#> [1] "    last loop bank: right descending"
#> [1] "    current loop bank: left descending"
#> [1] "        Bend 1"
#> [1] "            Start: TRUE Route-M: 286.12"
#> [1] "            End: TRUE   Route-M: 482.15"
#> [1] "            bank: left descending"
#> [1] "            min_loop_m: 286.12"
#> [1] "            max_loop_m: 482.15"
#> [1] "            min_last_bend_m: 286.12"
#> [1] "            max_last_bend_m: 482.15"
#> [1] "        Bend 2"
#> [1] "            Start: TRUE Route-M: 498.39"
#> [1] "            End: TRUE   Route-M: 720.69"
#> [1] "            bank: left descending"
#> [1] "            min_loop_m: 286.12"
#> [1] "            max_loop_m: 720.69"
#> [1] "            min_last_bend_m: 286.12"
#> [1] "            max_last_bend_m: 482.15"
#> [1] "            ** Bend 2 is upstream of Bend 1"
#> [1] "    ** The Loop 2 apex point is located within the loop."
#> [1] "    ** Loop 2 points are located on the opposite bank from Loop 1"
#> [1] "Loop 3"
#> [1] "    Apex: TRUE   Route-M: 998.08"
#> [1] "    last loop bank: left descending"
#> [1] "    current loop bank: right descending"
#> [1] "        Bend 1"
#> [1] "            Start: TRUE Route-M: 699.88"
#> [1] "            End: TRUE   Route-M: 888.34"
#> [1] "            bank: right descending"
#> [1] "            min_loop_m: 699.88"
#> [1] "            max_loop_m: 888.34"
#> [1] "            min_last_bend_m: 699.88"
#> [1] "            max_last_bend_m: 888.34"
#> [1] "        Bend 2"
#> [1] "            Start: TRUE Route-M: 900.55"
#> [1] "            End: TRUE   Route-M: 1248.2"
#> [1] "            bank: right descending"
#> [1] "            min_loop_m: 699.88"
#> [1] "            max_loop_m: 1248.2"
#> [1] "            min_last_bend_m: 699.88"
#> [1] "            max_last_bend_m: 888.34"
#> [1] "            ** Bend 2 is upstream of Bend 1"
#> [1] "    ** The Loop 3 apex point is located within the loop."
#> [1] "    ** Loop 3 points are located on the opposite bank from Loop 2"
#> [1] "Loop 4"
#> [1] "    Apex: TRUE   Route-M: 1920.21"
#> [1] "    last loop bank: right descending"
#> [1] "    current loop bank: left descending"
#> [1] "        Bend 1"
#> [1] "            Start: TRUE Route-M: 1217.34"
#> [1] "            End: TRUE   Route-M: 1755.37"
#> [1] "            bank: left descending"
#> [1] "            min_loop_m: 1217.34"
#> [1] "            max_loop_m: 1755.37"
#> [1] "            min_last_bend_m: 1217.34"
#> [1] "            max_last_bend_m: 1755.37"
#> [1] "        Bend 2"
#> [1] "            Start: TRUE Route-M: 1770.34"
#> [1] "            End: TRUE   Route-M: 2228.29"
#> [1] "            bank: left descending"
#> [1] "            min_loop_m: 1217.34"
#> [1] "            max_loop_m: 2228.29"
#> [1] "            min_last_bend_m: 1217.34"
#> [1] "            max_last_bend_m: 1755.37"
#> [1] "            ** Bend 2 is upstream of Bend 1"
#> [1] "    ** The Loop 4 apex point is located within the loop."
#> [1] "    ** Loop 4 points are located on the opposite bank from Loop 3"
#> [1] "Loop 5"
#> [1] "    Apex: TRUE   Route-M: 2629.73"
#> [1] "    last loop bank: left descending"
#> [1] "    current loop bank: right descending"
#> [1] "        Bend 1"
#> [1] "            Start: TRUE Route-M: 2325.57"
#> [1] "            End: TRUE   Route-M: 3048.42"
#> [1] "            bank: right descending"
#> [1] "            min_loop_m: 2325.57"
#> [1] "            max_loop_m: 3048.42"
#> [1] "            min_last_bend_m: 2325.57"
#> [1] "            max_last_bend_m: 3048.42"
#> [1] "    ** The Loop 5 apex point is located within the loop."
#> [1] "    ** Loop 5 points are located on the opposite bank from Loop 4"