Calculate shear stress for each cross section

Calculates shear stress for each cross section in the input xs_dims data frame.

shear_stress(xs_dims, water_density = 1000)

Arguments

xs_dims

data frame; A data frame of cross section dimensions.

water_density

numeric; density of water (units kg/m^3)

Value

Returns a data frame of cross sections with the calculated shear stress.

shear_stress_density Shear stress calculated using the density of water typically of the form "1000 D Sw". Units: \(kg/m^2\)

shear_stress_weight Shear stress calculated using the specific weight of water. Units: \(lb/ft^2\)

shear_stress_lane Shear stress calculated using just mean depth and water surface slope. Units:

Details

Shear Stress Equations Shear stress is a measure of the force of friction from a fluid acting on a body in the path of that fluid. In the case of open channel flow, it is the force of moving water against the bed of the channel. Generalized shear stress or tractive force is commonly defined using the following

Density of Water This functional form of the equation defines the \(\gamma\) term as the density of water.

$$\tau = \gamma D Sw$$

where: \(\tau\) is the fluid shear stress (\(kg/m^2\)), \(\gamma\) is the density of water (\(kg/m^3\)), \(D\) is mean water depth (m), \(Sw\) is the water surface slope (dimensionless: m/m).

Specific Weight of Water This functional form of the equation defines the \(\gamma\) term as the specific weight of water

$$\tau = \gamma D Sw$$

where: \(\tau\) is the fluid shear stress (\(lb/ft^2\)), \(\gamma\) is the specific weight of water (\(lb/ft^2\)), \(D\) is mean water depth (ft), \(Sw\) is the water surface slope (dimensionless: ft/ft).

where: \(\gamma = \rho a_{g}\)

\(\rho\) is the density of water (\(slugs/ft^3\)), \(a_{g}\) is the acceleration of gravity (\(ft/sec^2\))

Lane's Balance Version This functional form of the equation ignores absolute values and simply focuses on the variables that are changing:

$$\tau = D Sw$$

\(D\) is mean water depth (m), \(Sw\) is the water surface slope (dimensionless: m/m).

Typical Values

Density of water at \(4^{\circ} C (39^{\circ} F)\) = \(1000 kg/m^3\) or \(1.94032 slugs/ft^3\)

Acceleration of gravity = \(9.807 m/s^2, 32.174 ft/s^2\)

Examples

# Calculate cross section dimensions
xs_dims <- cross_section_dimensions_L2(xs = fluvgeo::sin_riffle_channel_sf,
                                       xs_points = fluvgeo::sin_riffle_channel_points_sf,
                                       bankfull_elevation = 103,
                                       lead_n = 1,
                                       use_smoothing = TRUE,
                                       loess_span = 0.5,
                                       vert_units = "ft")
#> [1] "Sinsinawa"
#> slope and sinuosity complete
#> xs 1 metrics complete
#> xs 2 metrics complete
#> xs 3 metrics complete
#> xs 4 metrics complete
#> xs 5 metrics complete
#> xs 6 metrics complete
#> xs 7 metrics complete
#> xs 8 metrics complete
#> xs 9 metrics complete
#> xs 10 metrics complete
#> join of reach_geoms and xs_reach_geoms complete

# Calculate shear stress
xs_dims_ss <- shear_stress(xs_dims)