Sinsinawa Creek, IL stream gradient points

A sf object containing stream gradient points for Sinsinawa Creek, IL. These points represent elevation values extracted from a LiDAR terrain surface along the flowline of the stream.

sin_gradient_sf

Format

A sf with 3521 observations and 23 variables.

OBJECTID

numeric; The ArcGIS feature class assigned unique identifier for each point in the dataset.

OBJECTID_1

numeric; The ArcGIS feature class assigned unique identifier for each point in the dataset.

ReachName

character; The name of the reach.

POINT_X

numeric; The longitude or easting of the point. Coordinate system units can be determined by checking the st_crs(sin_gradient_sf).

POINT_Y

numeric; The latitude or northing of the point. Coordinate system units can be determined by checking the st_crs(sin_gradient_sf).

POINT_M

numeric; The route distance of this cross section from the upstream end of the reach. Coordinate system units can be determined by checking the st_crs(sin_gradient_sf).

Z

numeric; The elevation of the point. By convention, the vertical units used in this project are NADV88 feet.

Z_smooth

numeric; The smoothed elevation of the point. By convention, the vertical unites used in this project are NAVD88 feet.

upstream_x

numeric; The x-coordinate value (longitude) of the next upstream point in the moving window.

upstream_y

numeric; The y-coordinate value (latitude) of the next upstream point in the moving window.

downstream_x

numeric; The x-coordinate value (longitude) of the next downsteam point in the moving window.

downstream_y

numeric; The y-coordinate value (latitude) of the next downsteam point in the moving window.

upstream_m

numeric; The m-coordinate value (route position) of the next upstream point in the moving window.

downstream_m

numeric; The m-coordinate value (route position) of the next downstream point in the moving window.

rise

numeric; The elevation difference between the downstream and upstream point in the moving window, units are in feet.

run

numeric; The longitudinal distance along the stream network between the downstream and upstream point in the moving window, units are in feet.

stream_length

numeric; The longitudinal distance along the stream network between the downstream and upstream point in the moving window, units are in feet.

valley_length

numeric; The longitudinal distance along the valleyline between the downstream and upstream stream network point in the moving window, units are in feet.

sinuosity

numeric; The sinuosity of the stream within the moving window (valley_length/stream_length).

sinuosity_gte_one

numeric;The sinuosity metric, values greater than or equal to one.

slope

numeric; The slope of the stream within the moving window (rise/run).

slope_gte_zero

numeric; The slope metric, values greater than of equal to zero.

Source

This dataset was produced using the FluvialGeomorph ArcGIS toolbox.