Sinsinawa Creek, IL stream riffle cross section lines

A sf object containing level 3 cross section dimensions for Sinsinawa Creek, IL. Cross section dimension polyline feature classes are used to store the hydraulic dimensions calculated for each cross section.

sin_riffle_floodplain_dims_L3_sf

Format

A sf with 10 observations and 74 variables.

OBJECTID

numeric; The ArcGIS feature class assigned unique identifier for each point in the dataset.

loop

numeric; The loop number.

bend

numeric; The bend number.

OBJECTID_1

numeric; The ArcGIS feature class assigned unique identifier for each point in the dataset.

Seq

numeric; The cross section unique identifier.

OBJECTID_12

numeric; The ArcGIS feature class assigned unique identifier for each point in the dataset.

Watershed_Area_SqMile

numeric; The upstream watershed drainage area.

from_measure

numeric; The start longitudinal stationing value for the reach in km.

to_measure

numeric; The end longitudinal stationing value for the reach in km.

ReachName

character; The name of the reach.

POINT_X

numeric; The x-coordinate value (longitude).

POINT_Y

numeric; The y-coordinate value (latitude).

POINT_M

numeric; The m-coordinate value (route position).

Z

numeric;The z-coordinate value (elevation).

km_to_mouth

numeric; The distance to the mouth of the site in km.

Z_smooth

numeric; The smoothed z-coordinate value (elevation).

upstream_x

numeric; The x-coordinate value (longitude) of the next upstream point in the moving window.

upstream_y

numeric; The y-coordinate value (latitude) of the next upstream point in the moving window.

downstream_x

numeric; The x-coordinate value (longitude) of the next downsteam point in the moving window.

downstream_y

numeric; The y-coordinate value (latitude) of the next downsteam point in the moving window.

upstream_z

numeric; The z-coordinate value (elevation) of the next upstream point in the moving window.

downstream_z

numeric; The z-coordinate value (elevation) of the next downstream point in the moving window.

upstream_m

numeric; The m-coordinate value (route position) of the next upstream point in the moving window.

downstream_m

numeric; The m-coordinate value (route position) of the next downstream point in the moving window.

rise

numeric; The elevation difference between the downstream and upstream point in the moving window, units are in feet.

run

numeric; The longitudinal distance along the stream network between the downstream and upstream point in the moving window, units are in feet.

stream_length

numeric; The longitudinal distance along the stream network between the downstream and upstream point in the moving window, units are in feet.

valley_length

numeric; The longitudinal distance along the valleyline between the downstream and upstream stream network point in the moving window, units are in feet.

sinuosity

numeric; The sinuosity of the stream within the moving window (valley_length/stream_length).

sinuosity_gte_one

numeric;The sinuosity metric, values greater than or equal to one.

slope

numeric; The slope of the stream within the moving window (rise/run).

slope_gte_zero

numeric; The slope metric, values greater than of equal to zero.

bankfull_elevation

numeric; The detrended bankfull elevation (in feet) that is used to calculate hydraulic geometry.

drainage_area

numeric;The drainage area (single value or vector) used to calculate discharge.

xs_area

numeric; The cross sectional area at the specified detrended bankfull elevation, units: square feet.

xs_width

numeric; The cross section width at the specified detrended bankfull elevation, units: feet.

xs_depth

numeric; The maximum depth at the specified detrended bankfull elevation, units: detrended feet.

discharge

numeric; The estimated discharge at the specified drainage area.

fp_area

numeric; The cross sectional area at the flood prone elevation, units: square feet.

fp_width

numeric; The cross section width at the flood prone elevation, units: feet.

fp_depth

numeric; The maximum depth at the flood prone elevation, units: feet.

xs_width_depth_ratio

numeric; The bankfull width to bankfull depth ratio.

xs_width_depth_ratio_gte_one

numeric; The bankfull width to bankfull depth ratio, greater than or equal to one.

xs_entrenchment_ratio

numeric; The entrenchment ratio (flood prone width) / (bankfull width).

xs_entrenchment_ratio_gte_one

numeric; The entrenchment ratio (flood prone width) / (bankfull width), greater than or equal to one.

watersurface_elev

numeric; The water surface elevation, units: NAVD88 feet.

bankfull_elev

numeric; The bankfull elevation, units: NAVD88 feet.

floodprone_elev

numeric; The flood prone elevation, units: NAVD88 feet.

xs_mean_depth

numeric; The cross section mean depth in feet at the estimated bankfull water surface elevation.

shear_stress_density

numeric; The fluid shear stress, calculated using the density of water.

shear_stress_density_gte_zero

numeric; The fluid shear stress density metric, greater than or equal to zero.

shear_stress_weight

numeric; The fluid shear stress, calculated using the specific weight of water.

shear_stress_weight_gte_zero

numeric; The fluid shear stress specific weight metric, greater than or equal to zero.

shear_stress_lane

numeric; The fluid shear stress, calculated using the Lane’s Balance method.

shear_stress_lane_gte_zero

numeric; The fluid shear stress Lane’s metric, greater than or equal to zero.

stream_power

numeric; The amount of energy the water is exerting on the channel.

stream_power_gte_zero

numeric; The stream power, greater than or equal to zero.

stream_power_lane

numeric; The stream power, calculated using the Lane's Balance method.

stream_power_lane_gte_zero

numeric; The fluid shear stress Lane’s metric, greater than or equal to zero.

unit_stream_power

numeric; The unit stream power.

unit_stream_power_gte_zero

numeric; The unit stream power metric, greater than or equal to zero.

bend_num

numeric; The bend number.

bend_POINT_X

numeric; The x-coordinate value (longitude) of the center of the bend

bend_POINT_Y

numeric; The y-coordinate value (latitude) of the center of the bend.

loop_POINT_X

numeric; The x-coordinate value (longitude) of the center of the loop.

loop_POINT_Y

numeric; The y-coordinate value (latitude) of the center of the loop.

bend_radius

numeric; The radius of the bend.

meander_length

numeric; The meander length in feet.

meander_width

numeric; The meander width in feet.

rc_bfw_ratio

numeric; The radius of curvature to bankfull width ratio.

rc_bfw_ratio_lte_30

numeric; The radius of curvature to bankfull width ratio less than or equal to 30.

mbw_bfw_ratio

numeric; The meander bend width to bankfull width ratio.

mbw_bfw_ratio_lte_30

numeric; The meander bend width to bankfull width ratio less than or equal to 30.

Source

This dataset was produced using the FluvialGeomorph ArcGIS toolbox.