Calculate stream power for each cross section

Calculates stream power for each cross section in the input xs_dims data frame.

stream_power(
  xs_dims,
  discharge_method = c("model_measure", "regional_curve", "width_relationship"),
  discharge_value = NULL,
  region = NULL,
  drainage_area = NULL,
  width_method = NULL
)

Arguments

xs_dims

data frame; A data frame of cross section dimensions.

discharge_method

character; The method for calculating discharge (Q). Must be one of: "model_measure", "regional_curve", "width_relationship".

discharge_value

numeric; The discharge value (single value or vector) to use for the stream power calculation. Required if discharge_method = "model_measure".

region

character; The regional curve name used to calculate discharge. Required if discharge_method = "regional_curve". This parameter is passed to the RegionalCurve::RHG function. See the RegionalCurve package for a list of regions with discharge relationships.

drainage_area

numeric; The drainage area (single value or vector) used by the RegionalCurve::RHG function to calculate discharge. Required if discharge_method = "regional_curve".

width_method

character; The name of the width relationship used to calculate discharge (Q) from width. Required if discharge_method = "width_relationship". Must be one of: <list goes here when implemented>

Value

Returns the input xs_dims data frame of cross sections with the calculated stream power variables added.

Details

Stream Power Equations Stream power is the rate of energy dissipation against the bed and banks of a river or stream per unit downstream length. It is given by the equation:

$$\Omega = \rho g Q S$$

where \(\Omega\) is the stream power (kg/m/sec), \(\rho\) is the density of water (assumes 1000 kg/m^3), \(g\) is acceleration due to gravity (9.8 m/s2), \(Q\) is discharge (m^3/s), and \(S\) is the channel slope (m/m).

Unit stream power is stream power per unit channel width, and is given by the equation:

$$\omega = (\rho g Q S) / b$$

where \(\omega\) is the unit stream power, and \(b\) is the width of the channel (m).

https://en.wikipedia.org/wiki/Stream_power

Lane's Balance Equation Stream power in the Lane's Balance equation is simply represented by discharge \(Q\) and slope \(S\):

$$\omega = Q S$$

Determining Discharge The discharge_method parameter specifies the source used for the discharge factor in the stream power equations.

model_measure Use this option to supply an estimate of discharge from a hydraulic model or field measurement.

regional_curve Use this option to use an estimate of discharge from a regional curve.

width_relationship Use this option to use an estimate of discharge from a width relationship

Examples

# Calculate cross section dimensions
xs_dims <- cross_section_dimensions_L2(xs = fluvgeo::sin_riffle_channel_sf,
                             xs_points = fluvgeo::sin_riffle_channel_points_sf,
                             bankfull_elevation = 103,
                             lead_n = 1,
                             use_smoothing = TRUE,
                             loess_span = 0.5,
                             vert_units = "ft")
#> [1] "Sinsinawa"
#> slope and sinuosity complete
#> xs 1 metrics complete
#> xs 2 metrics complete
#> xs 3 metrics complete
#> xs 4 metrics complete
#> xs 5 metrics complete
#> xs 6 metrics complete
#> xs 7 metrics complete
#> xs 8 metrics complete
#> xs 9 metrics complete
#> xs 10 metrics complete
#> join of reach_geoms and xs_reach_geoms complete

# Calculate stream power
xs_dims_sp <- stream_power(xs_dims,
                           discharge_method = "regional_curve",
                           region = "Illinois River",
                           drainage_area = 41)