Overplot multiple cross section profile surveys

Produces a cross section profile plot for the specified cross section for the input surveys.

xs_compare_plot_L2(
  stream,
  xs_number,
  xs_pts_sf_list,
  bankfull_elevation,
  aspect_ratio = 0.5,
  extent = "all"
)

Arguments

stream

character; The name of the stream.

xs_number

integer; The cross section `Seq` number.

xs_pts_sf_list

list; a list of `sf` objects of cross section points, one for each survey time period to be graphed. Survey list items must be tagged with the survey label to be used in the graph legend.

bankfull_elevation

numeric; The detrended bankfull elevation (in feet) that is used to calculate hydraulic geometry.

aspect_ratio

numeric; The aspect ratio of the graph.

extent

character; The extent of the cross section to plot. One of "all", "floodplain", or "channel".

Value

A ggplot2 object.

Details

This function is used to plot the cross section profile from a series of xs_points data frames representing multiple surveys.

See also