Calculate cross section and regional metrics for a set of streams, regions, and detrended bankfull elevations

Calculates a data frame of hydraulic geometry dimensions (area, width, depth) for a set of streams, regions, and detrended bankfull elevationsby cross section, xs type, and bankfull elevation.

xs_dimensions(xs_points, streams, regions, bankfull_elevations)

Arguments

xs_points

data frame; a data frame of cross section points

streams

character vector; The stream names in the study area.

regions

character; The regions that a dimension will be calculated for. See the regional_curves$region field for a complete list.

bankfull_elevations

numeric vector; The bankfull elevations (units: detrended feet) that are used to calculate hydraulic geometry.

Value

A data frame of cross section hydraulic geometry dimensions (area, width, depth, discharge) by cross section, xs type, and bankfull elevation.

reach_name

character; The name of the stream.

cross_section

numeric; The cross section unique identifier. Seq is only unique within a reach.

xs_type

character; A string indicating how the cross section was derived. "DEM derived cross section" denotes dimensions calculated from the DEM and "<Region Name>" denotes that the dimensions were calculated from that regions regional curve.

bankfull_elevation

numeric; The detrended bankfull elevation (in feet) that is used to calculate hydraulic geometry.

drainage_area

numeric; The area of the watershed upstream from this cross section, units: square miles.

xs_area

numeric; The cross sectional area at the specified detrended bankfull elevation, units: square feet.

xs_width

numeric; The cross section width at the specified detrended bankfull elevation, units: feet.

xs_depth

numeric; The maximum depth at the specified detrended bankfull elevation, units: detrended feet.

discharge

numeric; The estimated discharge at the specified drainage area.

See also

Examples

# Extract attribute data from the fluvgeo::sin_xs_points sf object
sin_xs_points_df <- fluvgeo::sin_riffle_channel_points_sf

# Call the xs_dimensions function with test data
streams <- c("Sinsinawa")
regions <- c("Eastern United States", "IN Central Till Plain")
bankfull_elevations = seq(103, 104, 0.1)

sin <- xs_dimensions(xs_points = sin_xs_points_df,
                     streams = streams,
                     regions = regions,
                     bankfull_elevations = bankfull_elevations)