Calculate cross section width, depth, area.

Calculates the hydraulic geometry (width, depth, area, and ground elevation) for the input cross section at the specified detrended elevation.

xs_geometry(xs_points, detrend_elevation)

Arguments

xs_points

data frame; a data frame of cross section points. Must be a single cross section.

detrend_elevation

numeric; The detrended elevation used to calculate hydraulic geometry, units: detrended feet.

Value

A data frame of hydraulic dimensions at the specified detrended elevation. The data frame contains the fields:

xs_width

numeric; The cross section width at the specified detrended elevation, units: feet.

xs_depth

numeric; The maximum depth at the specified detrended elevation, units: feet.

xs_area

numeric; The cross sectional area at the specified detrended elevation, units: square feet.

ground_elev

numeric; The ground elevation of the detrended elevation, units: NAVD88 feet.

Details

The cross section points used as input to this function must represent only one cross section. See the documentation for the example sin_xs_points data frame for the specification of this input sf::sf object.

See also

Examples

# Extract attribute data from the fluvgeo::sin_xs_points sf object
sin_xs_points_df <- fluvgeo::sin_riffle_channel_points_sf

# Subset sin_xs_points to contain only one cross section (Seq = 4)
sin_xs_points_4 <- sin_xs_points_df[sin_xs_points_df$Seq == 4,]

# Calculate hydraulic geometry for a single cross section
xs_geometry(xs_points = sin_xs_points_4, detrend_elevation = 103.5)
#>   xs_width xs_depth  xs_area discharge ground_elev
#> 1 93.32383 2.795617 259.4544         0    645.7752