Calculate cross section hydraulic geometry metrics

Calculates the hydraulic geometry metrics bankfull (width, depth, area, elevation), floodprone (width, depth, area, elevation), width-depth ratio, entrenchment ratio, and water surface elevation for the input cross section.

xs_metrics(xs_points, stream, xs_number, bankfull_elevation)

Arguments

xs_points

data frame; a data frame of cross section points

stream

character; The name of the stream.

xs_number

integer; The cross section identifier of the requested cross section.

bankfull_elevation

numeric; The detrended bankfull elevation (in feet) that is used to calculate hydraulic geometry.

Value

A data frame of hydraulic dimensions for the specified cross section at the specified detrended bankfull elevation.

reach_name

character; The name of the stream.

cross_section

numeric; The cross section unique identifier. Seq is only unique within a reach.

xs_type

character; A string indicating how the cross section was derived. "DEM derived cross section" denotes dimensions calculated from the DEM and "<Region Name>" denotes that the dimensions were calculated from that regions regional curve.

bankfull_elevation

numeric; The detrended bankfull elevation (in feet) that is used to calculate hydraulic geometry.

drainage_area

numeric; The area of the watershed upstream from this cross section, units: square miles.

xs_area

numeric; The cross sectional area at the specified detrended bankfull elevation, units: square feet.

xs_width

numeric; The cross section width at the specified detrended bankfull elevation, units: feet.

xs_depth

numeric; The maximum depth at the specified detrended bankfull elevation, units: detrended feet.

discharge

numeric; The estimated discharge at the specified drainage area.

fp_area

numeric; The cross sectional area at the flood prone elevation, units: square feet.

fp_width

numeric; The cross section width at the flood prone elevation, units: feet.

fp_depth

numeric; The maximum depth at the flood prone elevation, units: feet.

xs_width_depth_ratio

numeric; The bankfull width to bankfull depth ratio.

xs_entrenchment_ratio

numeric; The entrenchment ratio (flood prone width) / (bankfull width).

watersurface_elev

numeric; The water surface elevation, units: NAVD88 feet.

bankfull_elev

numeric; The bankfull elevation, units: NAVD88 feet.

floodprone_elev

numeric; The flood prone elevation, units: NAVD88 feet.

Details

Add methods description for calculating flood prone, width-depth ratio, and entrenchment ratio. Calls xs_geometry().

See also

Examples

# Extract attribute data from the fluvgeo::sin_riffle_floodplain_points_sf
# sf
sin_xs_points_df <- fluvgeo::sin_riffle_floodplain_points_sf

# Call the xs_metrics function
sin_4 <- xs_metrics(xs_points = sin_xs_points_df,
                    stream = "Sinsinawa",
                    xs_number = 4,
                    bankfull_elevation = 103.5)